backend/pyproject.toml
[project]
name = "google-analytics-backend"
version = "0.1.0"
description = "Google Analytics Intelligence System Backend Services"
authors = [
    {name = "Subramanya Nagabhushanaradhya", email = "subramanyanagabhushan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "pg8000>=1.30.3",
    "asyncpg>=0.29.0",
    "pandas>=2.1.3",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic[email]>=2.11.7",
    "httpx>=0.24.0",
    "loguru>=0.7.2",
    "pydantic-settings>=2.10.1",
    # Data service specific dependencies
    "tqdm>=4.66.1",
    "google-cloud-storage>=2.14.0",
    "google-auth>=2.23.4",
    "google-cloud-bigquery>=3.12.0",
    "db-dtypes>=1.1.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=23.2.0",
    "tenacity>=8.2.3",
    "paramiko>=3.4.0",
    "openpyxl>=3.1.2",
    "jinja2>=3.1.6",
    "aiohttp>=3.13.2",
    "azure-storage-queue>=12.14.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["common", "services"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "vulture>=2.14",
]
